Added functionality to scripts to optionally disable vagrant or delete vagrant accounts after setup, added functionality to create users script to make accounts from JSON files, made adding to local admin script more flexible

+++ scripts/setup-ad/remove-vagrant.ps1
@@ -1,3 +1,15 @@
-Remove-LocalUser vagrant
+param(
+    [string] $disableOrDel = "disable"
+)
 
-Remove-ADUser -Identity vagrant
+if($disableOrDel -eq "disable") {
+    Disable-LocalUser -Name "vagrant"
+    if (Get-Module -ListAvailable -Name ActiveDirectory) {
+        Disable-ADAccount -Identity vagrant
+    }
+} elseif ($disableOrDel -eq "delete") {
+    Remove-LocalUser vagrant
+    if (Get-Module -ListAvailable -Name ActiveDirectory) {
+        Remove-ADUser -Identity vagrant
+    }
+}
